// src/game-server/Enums/Skill.ts
export const enum Skill {
	FrailtyI = 0,
	FrailtyII = 1,
	FrailtyIII = 2,
	FrailtyIV = 3,
	PoisonI = 4,
	PoisonII = 5,
	PoisonIII = 6,
	PoisonIV = 7,
	ChaosI = 8,
	ChaosII = 9,
	ChaosIII = 10,
	ChaosIV = 11,
	HypnotizeI = 12,
	HypnotizeII = 13,
	HypnotizeIII = 14,
	HypnotizeIV = 15,
	StunI = 16,
	StunII = 17,
	StunIII = 18,
	StunIV = 19,
	PurgeChaosI = 20,
	PurgeChaosII = 21,
	PurgeChaosIII = 22,
	PurgeChaosIV = 23,
	UnStunI = 24,
	UnStunII = 25,
	UnStunIII = 26,
	UnStunIV = 27,
	MultiShotI = 28,
	MultiShotII = 29,
	MultiShotIII = 30,
	MultiShotIV = 31,
	BlizzardI = 32,
	BlizzardII = 33,
	BlizzardIII = 34,
	BlizzardIV = 35,
	SpeedI = 36,
	SpeedII = 37,
	SpeedIII = 38,
	SpeedIV = 39,
	HealOtherI = 40,
	HealOtherII = 41,
	HealOtherIII = 42,
	HealOtherIV = 43,
	FireI = 44,
	FireII = 45,
	FireIII = 46,
	FireIV = 47,
	IceI = 48,
	IceII = 49,
	IceIII = 50,
	IceIV = 51,
	EvilI = 52,
	EvilII = 53,
	EvilIII = 54,
	EvilIV = 55,
	FlashI = 56,
	FlashII = 57,
	FlashIII = 58,
	FlashIV = 59,
	DeathI = 60,
	DeathII = 61,
	DeathIII = 62,
	DeathIV = 63,
	EnhanceI = 64,
	EnhanceII = 65,
	EnhanceIII = 66,
	EnhanceIV = 67,
	ProtectI = 68,
	ProtectII = 69,
	ProtectIII = 70,
	ProtectIV = 71,
	DrainI = 72,
	DrainII = 73,
	DrainIII = 74,
	DrainIV = 75,
	ReflectI = 76,
	ReflectII = 77,
	ReflectIII = 78,
	ReflectIV = 79,
	RepelI = 80,
	RepelII = 81,
	RepelIII = 82,
	RepelIV = 83,
}